Updated settings dialog, moved it to it's own file
