+++ settings.ui
@@ -197,10 +197,10 @@ color: rgb(211, 215, 207);</string>
     <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
    </property>
   </widget>
-  <widget class="QCheckBox" name="useqrz_checkBox">
+  <widget class="QRadioButton" name="useqrz_radioButton">
    <property name="geometry">
     <rect>
-     <x>130</x>
+     <x>62</x>
      <y>140</y>
      <width>101</width>
      <height>25</height>
@@ -218,11 +218,14 @@ color: rgb(211, 215, 207);</string>
     <bool>false</bool>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>Use QRZ</string>
    </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup_2</string>
+   </attribute>
   </widget>
   <widget class="QCheckBox" name="usecloudlog_checkBox">
    <property name="geometry">
@@ -242,16 +245,16 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>Use CloudLog</string>
    </property>
   </widget>
-  <widget class="QCheckBox" name="userigcontrol_checkBox">
+  <widget class="QRadioButton" name="userigctld_radioButton">
    <property name="geometry">
     <rect>
-     <x>190</x>
+     <x>140</x>
      <y>420</y>
      <width>101</width>
      <height>25</height>
@@ -266,11 +269,14 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>rigctld</string>
    </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup</string>
+   </attribute>
   </widget>
   <widget class="QLineEdit" name="lookup_user_name_field">
    <property name="geometry">
@@ -462,9 +468,9 @@ color: rgb(211, 215, 207);</string>
   <widget class="QLabel" name="label_8">
    <property name="geometry">
     <rect>
-     <x>20</x>
+     <x>10</x>
      <y>486</y>
-     <width>137</width>
+     <width>151</width>
      <height>25</height>
     </rect>
    </property>
@@ -475,7 +481,10 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="text">
-    <string>XPlanet Marker File:</string>
+    <string>XPlanet Marker:</string>
+   </property>
+   <property name="alignment">
+    <set>Qt::AlignRight|Qt::AlignTrailing|Qt::AlignVCenter</set>
    </property>
   </widget>
   <widget class="QLineEdit" name="markerfile_field">
@@ -519,10 +528,10 @@ color: rgb(211, 215, 207);</string>
     <string>Generate Marker</string>
    </property>
   </widget>
-  <widget class="QCheckBox" name="usehamdb_checkBox">
+  <widget class="QRadioButton" name="usehamdb_radioButton">
    <property name="geometry">
     <rect>
-     <x>250</x>
+     <x>182</x>
      <y>140</y>
      <width>121</width>
      <height>25</height>
@@ -535,16 +544,19 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>Use HamDB</string>
    </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup_2</string>
+   </attribute>
   </widget>
-  <widget class="QCheckBox" name="checkBox">
+  <widget class="QRadioButton" name="usehamqth_radioButton">
    <property name="geometry">
     <rect>
-     <x>390</x>
+     <x>322</x>
      <y>140</y>
      <width>121</width>
      <height>23</height>
@@ -557,11 +569,14 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>Use HamQTH</string>
    </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup_2</string>
+   </attribute>
   </widget>
   <widget class="QLabel" name="label_3">
    <property name="geometry">
@@ -585,10 +600,10 @@ color: rgb(211, 215, 207);</string>
     <set>Qt::AlignCenter</set>
    </property>
   </widget>
-  <widget class="QCheckBox" name="checkBox_2">
+  <widget class="QRadioButton" name="useflrig_radioButton">
    <property name="geometry">
     <rect>
-     <x>320</x>
+     <x>270</x>
      <y>420</y>
      <width>92</width>
      <height>23</height>
@@ -601,11 +616,14 @@ color: rgb(211, 215, 207);</string>
     </font>
    </property>
    <property name="styleSheet">
-    <string notr="true">border: 1px solid red;</string>
+    <string notr="true"/>
    </property>
    <property name="text">
     <string>flrig</string>
    </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup</string>
+   </attribute>
   </widget>
   <widget class="QLabel" name="label_9">
    <property name="geometry">
@@ -628,6 +646,50 @@ color: rgb(211, 215, 207);</string>
    <property name="alignment">
     <set>Qt::AlignCenter</set>
    </property>
+  </widget>
+  <widget class="QRadioButton" name="radioButton">
+   <property name="geometry">
+    <rect>
+     <x>390</x>
+     <y>420</y>
+     <width>71</width>
+     <height>23</height>
+    </rect>
+   </property>
+   <property name="font">
+    <font>
+     <family>JetBrains Mono</family>
+     <pointsize>12</pointsize>
+    </font>
+   </property>
+   <property name="text">
+    <string>None</string>
+   </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup</string>
+   </attribute>
+  </widget>
+  <widget class="QRadioButton" name="radioButton_3">
+   <property name="geometry">
+    <rect>
+     <x>462</x>
+     <y>140</y>
+     <width>71</width>
+     <height>23</height>
+    </rect>
+   </property>
+   <property name="font">
+    <font>
+     <family>JetBrains Mono</family>
+     <pointsize>12</pointsize>
+    </font>
+   </property>
+   <property name="text">
+    <string>None</string>
+   </property>
+   <attribute name="buttonGroup">
+    <string notr="true">buttonGroup_2</string>
+   </attribute>
   </widget>
  </widget>
  <resources/>
@@ -665,4 +727,8 @@ color: rgb(211, 215, 207);</string>
    </hints>
   </connection>
  </connections>
+ <buttongroups>
+  <buttongroup name="buttonGroup"/>
+  <buttongroup name="buttonGroup_2"/>
+ </buttongroups>
 </ui>
